# split-bill-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.BankingDeeplink:
    properties:
      app_name:
        type: string
      color:
        type: string
      icon_name:
        type: string
      scheme:
        type: string
    type: object
...
  models.PaymentDeeplinkRequest:
    properties:
      account_name:
        type: string
      account_number:
        type: string
      amount:
        type: number
      bank_code:
        type: string
      note:
        type: string
    required:
    - account_number
    - amount
    - bank_code
    type: object
  models.PaymentDeeplinkResponse:
    properties:
      amount:
        type: number
      deeplinks:
        items:
          $ref: '#/definitions/models.BankingDeeplink'
        type: array
      note:
        type: string
      vietqr_url:
        type: string
    type: object
...
  models.VietQRRequest:
    properties:
      account_name:
        type: string
      account_no:
        type: string
      amount:
        type: number
      bank_id:
        type: string
      description:
        type: string
      template:
        type: string
    required:
    - account_no
    - amount
    - bank_id
    type: object
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@splitbill.app
    name: Split Bill Support
  description: API server for Split Bill application - manage groups, bills, transactions,
    OCR receipt scanning, and payments
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Split Bill API
  version: 4.0.0
paths:
...
  /groups/{id}/export:
    get:
      description: Generates a text or JSON summary of a group's bills and settlements
        for sharing
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Export format: text or json (default text)'
        in: query
        name: format
        type: string
      produces:
      - application/json
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Export group summary
      tags:
      - Stats
...
  /groups/{id}/stats:
    get:
      description: Returns comprehensive statistics for a group including spending,
        categories, and trends
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get group statistics
      tags:
      - Stats
  /groups/{id}/stats/categories:
    get:
      description: Returns category spending breakdown and monthly trends for a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get group category statistics
      tags:
      - Stats
...
  /payment/banks:
    get:
      description: Returns the list of supported Vietnamese banks with IDs, names,
        codes, logos, and colors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get supported banks
      tags:
      - Payment
  /payment/deeplink:
    post:
      consumes:
      - application/json
      description: Generates banking app deeplinks (Momo, ZaloPay, VNPay, etc.) and
        VietQR URL for payment
      parameters:
      - description: Payment details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PaymentDeeplinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.PaymentDeeplinkResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Generate payment deeplinks
      tags:
      - Payment
  /payment/user/{userId}:
    get:
      description: Returns payment information (bank accounts, preferred payment method)
        for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user payment info
      tags:
      - Payment
  /payment/vietqr:
    post:
      consumes:
      - application/json
      description: Generates a VietQR code image URL for bank transfer payment
      parameters:
      - description: QR code details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.VietQRRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Generate VietQR code
      tags:
      - Payment
  /stats/me:
    get:
      description: Returns overall statistics for the authenticated user across all
        groups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get current user's statistics
      tags:
      - Stats
...
securityDefinitions:
  BearerAuth:
    description: 'Firebase Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"